Rewrite estimation_table test
