+++ tests/visualization/test_estimation_table.py
@@ -1,3 +1,6 @@
+import io
+import textwrap
+
 import numpy as np
 import pandas as pd
 import pytest
@@ -26,17 +29,59 @@ from estimagic.visualization.estimation_table import render_latex
 from pandas.testing import assert_frame_equal as afe
 from pandas.testing import assert_series_equal as ase
 
-from tests.visualization.helpers_test_estimation_table import (
-    _get_models_multiindex,
-)
-from tests.visualization.helpers_test_estimation_table import (
-    _get_models_multiindex_multi_column,
-)
-from tests.visualization.helpers_test_estimation_table import (
-    _get_models_single_index,
-)
-from tests.visualization.helpers_test_estimation_table import _read_csv_string
 
+# ======================================================================================
+# Helper functions
+# ======================================================================================
+def _get_models_multiindex():
+    df = pd.DataFrame(
+        data=np.ones((3, 4)), columns=["value", "ci_lower", "ci_upper", "p_value"]
+    )
+    df.index = pd.MultiIndex.from_tuples(
+        [("p_1", "v_1"), ("p_1", "v_2"), ("p_2", "v_2")]
+    )
+    info = {"n_obs": 400}
+    mod1 = {"params": df, "info": info, "name": "m1"}
+    mod2 = {"params": df, "info": info, "name": "m2"}
+    models = [mod1, mod2]
+    return models
+
+
+def _get_models_single_index():
+    df = pd.DataFrame(
+        data=np.ones((3, 4)), columns=["value", "ci_lower", "ci_upper", "p_value"]
+    )
+    df.index = [f"p{i}" for i in [1, 2, 3]]
+    info = {"n_obs": 400}
+    mod1 = {"params": df, "info": info, "name": "m1"}
+    mod2 = {"params": df, "info": info, "name": "m2"}
+    models = [mod1, mod2]
+    return models
+
+
+def _get_models_multiindex_multi_column():
+    df = pd.DataFrame(
+        data=np.ones((3, 4)), columns=["value", "ci_lower", "ci_upper", "p_value"]
+    )
+    df.index = pd.MultiIndex.from_tuples(
+        [("p_1", "v_1"), ("p_1", "v_2"), ("p_2", "v_2")]
+    )
+    info = {"n_obs": 400}
+    mod1 = {"params": df.iloc[1:], "info": info, "name": "m1"}
+    mod2 = {"params": df, "info": info, "name": "m2"}
+    mod3 = {"params": df, "info": info, "name": "m2"}
+    models = [mod1, mod2, mod3]
+    return models
+
+
+def _read_csv_string(string, index_cols=None):
+    string = textwrap.dedent(string)
+    return pd.read_csv(io.StringIO(string), index_col=index_cols)
+
+
+# ======================================================================================
+# Tests
+# ======================================================================================
 
 # test process_model for different model types
